% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{assignIntervalCalls}
\alias{assignIntervalCalls}
\title{Given a set of segments with calls, intersect the capture intervals with these segments and assign
calls to them.}
\usage{
assignIntervalCalls(intervals, gsegments, normalized_tumor_coverage)
}
\arguments{
\item{intervals}{a dataframe with a definition of the intervals used in the capture experiment}

\item{gsegments}{dataframe returned from the callSegments() function.}

\item{normalized_tumor_coverage}{a vector returned from robustNorm() function.}
}
\value{
a dataframe combining intervals and calls
}
\description{
Given a set of segments with calls, intersect the capture intervals with these segments and assign
calls to them.
}

