% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{callSegments}
\alias{callSegments}
\title{Make calls on segments based on a cutoff calculated from segment SD and a tuning factor.}
\usage{
callSegments(gsegments, normCoverage, tuning = 0.8)
}
\arguments{
\item{gsegments}{a dataframe with columns: Sample,Chromosome,Start,End,NumProbes,Segment,SegmentMean}

\item{normCoverage}{a vector or dataframe containing normalized coverage}

\item{tuning}{a parameter for adjusting sensitivity, higher values are less sensitive, more specific.}
}
\value{
A new dataframe with a "SegmentCall" column consisting of (-, 0, +).
}
\description{
Make calls on segments based on a cutoff calculated from segment SD and a tuning factor.
}
\examples{
\dontrun{
callSegments(gsegments, normCoverage)
callSegments(gsegments, normCoverage, tuning=.2)
}
}

