% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{formatForNexus}
\alias{formatForNexus}
\title{Create an output formatted for Nexus compatibility.}
\usage{
formatForNexus(perIntervalCalls)
}
\arguments{
\item{perIntervalCalls}{a data.frame containing per-interval calls}
}
\value{
A data.frame formatted for Nexus compatibility
}
\description{
Create an output formatted for Nexus compatibility.
}
\examples{
\dontrun{
formatForNexus(perIntervalCalls)
}
}

