% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plotGenomeIntervalCalls}
\alias{plotGenomeIntervalCalls}
\title{Make genome plots.}
\usage{
plotGenomeIntervalCalls(perIntervalCalls, main = "")
}
\arguments{
\item{perIntervalCalls}{data.frame containing per-Interval calls and other data.}

\item{main}{title of plot.}
}
\value{
nothing.
}
\description{
Make genome plots.
}
\examples{
\dontrun{
plotGenomeIntervalCalls(perIntervalCalls)
}
}

