% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plotSampleClusters}
\alias{plotSampleClusters}
\title{Make sample-cluster plots.}
\usage{
plotSampleClusters(rawNormalCoverage, rawTumorCoverage)
}
\arguments{
\item{rawNormalCoverage}{Data.frame containing raw coverage for normal samples}

\item{rawTumorCoverage}{Data.frame containing raw coverage for tumor samples}
}
\value{
nothing.
}
\description{
Make sample-cluster plots.
}
\examples{
\dontrun{
plotSampleClusters(rawNormalCoverage, rawTumorCoverage)
}
}

