% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{robustNorm}
\alias{robustNorm}
\title{Normalize coverage data.}
\usage{
robustNorm(y, x, pGC, EVs = NA, idx = NA)
}
\arguments{
\item{y}{a vector of coverage values for the sample of interest.}

\item{x}{a dataframe or matrix with coverage values for normal samples (columns are samples).}

\item{pGC,}{the percent GC content for each interval}

\item{EVs}{a vector or dataframe of explanatory variables (e.g. interval lengths).}

\item{idx}{a vector used to subset the input during model generation only (all intervals are reported)}
}
\value{
A new vector normalized against the panel of normals, interval lengths, and for GC content
}
\description{
Do a normalization with interval length, gcContent and panel of normals using a robust regression strategy
followed by a GC correction using loess.
}
\examples{
\dontrun{
robustNorm(y, x, ilengths, pGC)
}
}

