% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{segment}
\alias{segment}
\title{Segment normalized log2 coverage values for one or more samples.}
\usage{
segment(normCoverage, orderedChrom, maploc, idx = NA)
}
\arguments{
\item{normCoverage}{a data.frame or matrix with normalized coverage. Columns should be samples.}

\item{orderedChrom}{a vector of ordered factors indicating the chromosome, must be nrow(norm.coverage) long.}

\item{maploc}{a vector containing chromosome locus for intervals (end of each interval)}

\item{idx}{segments can be called using a subset of targeted intervals, idx provides the subset.}
}
\value{
A new dataframe with columns: Sample, Chromosome, Start, End, NumProbes, SegmentMean, SegmentCall
}
\description{
Segment normalized log2 coverage values for one or more samples.
}
\examples{
\dontrun{
segment(normalizedTumorCoverage, orderedChrom, maploc)
}
}

