% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{summarizeIntervalCalls}
\alias{summarizeIntervalCalls}
\title{Summarize interval level calls to gene level call. This is an interval function that
is used by summarizeToGeneCalls.}
\usage{
summarizeIntervalCalls(x)
}
\arguments{
\item{x}{a vector of calls for a single gene}
}
\value{
A data.frame containing QC metrics
}
\description{
Summarize interval level calls to gene level call. This is an interval function that
is used by summarizeToGeneCalls.
}
\examples{
\dontrun{
CNVqc(data)
}
}

