% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{summarizeToGeneCalls}
\alias{summarizeToGeneCalls}
\title{Summarize per-IntervalCalls to gene level calls.}
\usage{
summarizeToGeneCalls(perIntervalCalls)
}
\arguments{
\item{perIntervalCalls}{a vector of calls for a single gene}
}
\value{
A data.frame containing calls summarized to gene level
}
\description{
Summarize per-IntervalCalls to gene level calls.
}
\examples{
\dontrun{
CNVqc(data)
}
}

